server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Driver Tracker API Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.0",
    "djangorestframework>=3.15.0",
    "django-cors-headers>=4.4.0",
    "drf-spectacular>=0.27.2",
    "djangorestframework-simplejwt>=5.3.1",
    "drf-nested-routers>=0.93.5",
]

[project.scripts]
manage = "manage:main"
dev = "tasks:dev"
migrate = "tasks:migrate"
makemigrations = "tasks:makemigrations"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
